+++ src/main/java/org/openlmis/fulfillment/web/ProofOfDeliveryController.java
@@ -57,6 +57,8 @@ import org.slf4j.profiler.Profiler;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.oauth2.provider.OAuth2Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.Transactional;
@@ -66,6 +68,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.servlet.view.jasperreports.JasperReportsMultiFormatView;
 
 import java.util.List;
@@ -312,6 +315,47 @@ public class ProofOfDeliveryController extends BaseController {
 
     profiler.stop().log();
     XLOGGER.exit();
+  }
+
+  /**
+   * Get the audit information related to the given proof of delivery.
+   *
+   * @param author              The author of the changes which should be returned.
+   *                            If null or empty, changes are returned regardless of author.
+   * @param changedPropertyName The name of the property about which changes should be returned.
+   *                            If null or empty, changes associated with any and all properties
+   *                            are returned.
+   * @param page                A Pageable object that allows client to optionally add "page"
+   *                            (page number) and "size" (page size) query parameters to the
+   *                            request.
+   */
+  @RequestMapping(value = "proofsOfDelivery/{id}/auditLog", method = RequestMethod.GET)
+  @ResponseStatus(HttpStatus.OK)
+  @ResponseBody
+  public ResponseEntity<String> getAuditLog(
+      @PathVariable("id") UUID id,
+      @RequestParam(name = "author", required = false, defaultValue = "") String author,
+      @RequestParam(name = "changedPropertyName", required = false, defaultValue = "")
+          String changedPropertyName,
+      //Because JSON is all we formally support, returnJSON is excluded from our JavaDoc
+      @RequestParam(name = "returnJSON", required = false, defaultValue = "true")
+          boolean returnJson,
+      Pageable page,
+      OAuth2Authentication authentication) {
+    Profiler profiler = new Profiler("GET_AUDIT_LOG");
+    profiler.setLogger(LOGGER);
+
+    ProofOfDelivery pod = findProofOfDelivery(id, profiler);
+    canManagePod(authentication, profiler, pod);
+
+    profiler.start("GET_AUDIT_LOG");
+    ResponseEntity<String> response = getAuditLogResponse(
+        ProofOfDelivery.class, id, author, changedPropertyName, page, returnJson
+    );
+
+    profiler.stop().log();
+
+    return response;
   }
 
   private ProofOfDelivery findProofOfDelivery(UUID id, Profiler profiler) {